+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Karafka framework changelog
 
-## 2.0.41 (Unreleased)
+## 2.0.41 (2023-14-19)
 - **[Feature]** Provide `Karafka::Pro::Iterator` for anonymous topic/partitions iterations and messages lookups (#1389 and #1427).
 - [Improvement] Optimize topic lookup for `read_topic` admin method usage.
 - [Improvement] Report via `LoggerListener` information about the partition on which a given job has started and finished.